+++ CHANGELOG.md
@@ -17,6 +17,15 @@ _None_
 * Swift 4 templates.  
   [David Jennes](https://github.com/djbe) 
   [#67](https://github.com/SwiftGen/templates/pull/67)
+* XCAssets: Added support for named colors.  
+  [David Jennes](https://github.com/djbe) 
+  [#68](https://github.com/SwiftGen/templates/pull/68)
+* Fonts: the path to fonts will now default to just the font filename, but you can disable this behaviour by enabling the `preservePath` parameter.  
+  [David Jennes](https://github.com/djbe) 
+  [#71](https://github.com/SwiftGen/templates/pull/71)
+* Colors: new template that uses `#colorLiteral`s.  
+  [David Jennes](https://github.com/djbe) 
+  [#72](https://github.com/SwiftGen/templates/pull/72)
 
 ### Internal Changes